requirements.txt
pytest~=7.2.1
selene~=2.0.0b17
allure-pytest